% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmacs.R
\name{es_lavaan}
\alias{es_lavaan}
\alias{pin_effsize}
\title{Item-level effect size for non-invariance}
\usage{
es_lavaan(object, ...)

pin_effsize(object, ...)
}
\arguments{
\item{object}{A CFA model of class \code{\link[lavaan:lavaan-class]{lavaan::lavaan}}
fitted by \code{\link[lavaan:cfa]{lavaan::cfa()}}}

\item{...}{Additional arguments passed to \code{\link[=dmacs]{dmacs()}} or \code{\link[=fmacs]{fmacs()}}}
}
\value{
A matrix of 1 row showing the effect size values for
each non-invariant item on each latent variable.
}
\description{
For two groups, the function uses \code{\link[=dmacs]{dmacs()}} to compute
\eqn{d_\text{MACS}}. For more than two groups, the function
uses \code{\link[=fmacs]{fmacs()}} to compute \eqn{f_\text{MACS}}, a
generalisation of \eqn{d_\text{MACS}} similar to the
Cohen's \eqn{f} effect size.
}
