% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lui_sim}
\alias{lui_sim}
\title{Simulated data with 15 items of an Alcohol Belief Scale}
\format{
\subsection{\code{lui_sim}}{

A data frame with 910 rows and 16 columns:
\describe{
\item{class1 to class15}{Binary items}
\item{group}{six groups: 1 = White Male, 2 = Asian Male,
3 = Hispanic Male, 4 = White Female,
5 = Asian Female, 6 = Hispanic Female}

}
}
}
\source{
\url{https://osf.io/wxjsg}
}
\usage{
lui_sim
}
\description{
Simulation based on fitting a configural invariance model to
data made public by Lui (2019,
\url{https://psycnet.apa.org/record/2019-08221-001}). The original data
can be found at \url{https://osf.io/wxjsg}.
}
\keyword{datasets}
