% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinSearch.R
\name{pinSearch}
\alias{pinSearch}
\title{Search for invariant loadings/intercepts.}
\usage{
pinSearch(
  config_mod,
  data = NULL,
  group = NULL,
  ...,
  type = c("loadings", "intercepts", "residuals"),
  sig_level = 0.05
)
}
\arguments{
\item{config_mod}{Syntax of a configural invariance model to be passed to
\code{\link[lavaan]{cfa}}.}

\item{data}{A data frame to be passed to \code{\link[lavaan]{cfa}}.}

\item{group}{Character indicating the variable name in \code{data} that
defines the grouping variable in multiple-group CFA.}

\item{...}{Additonal arguments passed to \code{\link[lavaan]{cfa}}.}

\item{type}{Character variable indicating the stage of invariance to be
searched.}

\item{sig_level}{Significance level used to determine whether the parameter
associated with the highest modification index should be removed. Default
is .05.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Search for invariant loadings/intercepts.
}
\examples{
library(lavaan)
library(MASS)
# Simulate random data
set.seed(14)  # for reproducible results
mod_ninv <- "f =~ c(1, 0.8, .6)*y1 + c(0.8, 1.2, 1.0)*y2"
mod_inv <- paste0("1*y", 3:6, collapse = " + ")
sim_mod <- paste(
    paste(mod_ninv, "+", mod_inv),
    "f ~~ c(1, 1.3, 1.5)*f
     f ~ c(0, 0.5, 1.0)*1
     y1 ~ c(0, 0.3, 0)*1
     y3 ~ c(0.3, 0, -0.3)*1
     y1 ~~ c(1, .5, 1)*y1",
    sep = "\n"
)
# The uniqueness of each item is assumed to be 1.0
dat_sim = simulateData(sim_mod, sample.nobs = c(100, 100, 100))
# Fit configural model:
sam_config <- paste(
    paste0('f =~ ', paste0("y", 1:6, collapse = " + "))
)
pinSearch(sam_config, data = dat_sim, group = "group",
          type = "intercepts")
}
